This is a large commit, including a first draft of light pattern effects, timer synchronization and TWI communication.
First, light patterns have been implemented to realize some effects used in debugging the rest of the program. Ultimately, these effect patterns will need to be parameterized and controlled more granularly with TWI commands.

Second, a TWI event handler was implemented to consume bus commands. The current schema expects all units to be of a single physcial TWI address, but actually commanded via the first byte of this address, which is masked by the individual device's position on the aircraft.

Finally, a timer synchronization scheme was implemnted to adjust the pattern timer via phase correction signals received on the TWI general call address.

In addition, many routines were moved to the mainloop, avoiding any extraneous processing cycles in the ISRs. This activity is not completed.